bitmap: Use size_t instead of unsigned long
Use size_t instead of unsigned long to be consistent with the other
methods.

This is an omission from Ifb24da16b3282b72a610dc8f6f6c98f37623edd5.

+++ src/lib/libk/bitmap.cpp
@@ -322,7 +322,7 @@ public:
    *
    * \param bit  Bit index to clear.
    */
-  void atomic_clear_bit(unsigned long bit)
+  void atomic_clear_bit(size_t bit)
   {
     size_t idx = bit / Bpl;
     size_t pos = bit % Bpl;
